Add support for latest version of packages, better handling when no internet connection

+++ creepclient/repository.py
@@ -2,6 +2,7 @@
 
 import json # JSON encoder and decoder
 import os # Miscellaneous operating system interfaces
+import re # Regular expressions
 
 from operator import attrgetter
 from entity.package import Package
@@ -13,7 +14,11 @@ class Repository(object):
     version_hash = ''
     version_date = ''
 
+    # Has every package, including every version of each
     packages = []
+
+    # List of all packages, but only the latest version
+    unique_packages = []
 
     def __init__(self, appdir):
         self.localdir = appdir + os.sep + 'packages.json'
@@ -21,17 +26,24 @@ class Repository(object):
     def download_remote_repository(self):
         import urllib2
 
-        response = urllib2.urlopen(self.remote_url)
+        try:
+            response = urllib2.urlopen(self.remote_url)
+        except urllib2.URLError:
+            return False
+
         data = response.read()
 
         f = open(self.localdir, 'w')
         f.write(data)
         f.close()
+        return True
 
     def load_repository(self):
         # Repository file doesn't exist, fetch it from remote url
         if not os.path.isfile(self.localdir):
-            self.download_remote_repository()
+            if not self.download_remote_repository():
+                print "Package definition file not found or no internet connection."
+                return {'packages': {}}
             return json.load(open(self.localdir))
         
         # Check repository file date last modified
@@ -40,12 +52,14 @@ class Repository(object):
         import time
         filetime = os.stat(self.localdir).st_mtime
         if filetime + 3600 < calendar.timegm(time.gmtime()):
-            self.download_remote_repository()
+            if not self.download_remote_repository():
+                print "No internet connection. Using current version of repository. Date: {}".format(filetime)
 
         return json.load(open(self.localdir))
 
     def clear_cache(self):
-        os.remove(self.localdir)
+        if os.path.isfile(self.localdir):
+            os.remove(self.localdir)
 
     def populate(self, location=''):
         if not location:
@@ -80,6 +94,38 @@ class Repository(object):
                 self.packages.append(package)
 
         self.packages.sort(key=attrgetter('name'))
+        self.reduce_to_unique_packages()
+
+    def reduce_to_unique_packages(self):
+        """Make a listing of packages with only the latest version for each one"""
+        package_dict = {};
+        self.unique_packages = []
+
+        # First make a dict of packages index by name
+        for package in self.packages:
+            if not package.name in package_dict:
+                package_dict[package.name] = []
+            package_dict[package.name].append(package)
+
+        # Now go through dict and find the latest version of each
+        for name in package_dict:
+            packages = package_dict[name]
+            if len(packages) == 1:
+                self.unique_packages.append(packages[0])
+            else:
+                latest = packages[0]
+                for package in packages:
+                    if self.compare_versions(package.version, latest.version) > 0:
+                        latest = package
+                self.unique_packages.append(latest)
+
+        self.unique_packages.sort(key=attrgetter('name'))
+
+    def compare_versions(self, version1, version2):
+        return cmp(self.normalize_version(version1), self.normalize_version(version2))
+
+    def normalize_version(self, v):
+        return [int(x) for x in re.sub(r'(\.0+)*$','', v).split(".")]
 
     def count_packages(self):
         return len(self.packages)
@@ -88,7 +134,9 @@ class Repository(object):
         if name == '':
             return False
 
-        for package in self.packages:
+        # only select from the latest versions (unique_packages)
+        # TODO allow to fetch a package for a specific version
+        for package in self.unique_packages:
             if package.name == name:
                 return package
 
@@ -103,7 +151,7 @@ class Repository(object):
 
     def search(self, term):
         results = []
-        for package in self.packages:
+        for package in self.unique_packages:
             if term in package.name.split('/'):
                 results.append(package)
                 continue
